#893 validate posts with serializer before deletion

+++ misago/threads/api/postendpoints/delete.py
@@ -9,6 +9,7 @@ from misago.core.utils import clean_ids_list
 from misago.threads.moderation import posts as moderation
 from misago.threads.permissions import allow_delete_event, allow_delete_post
 from misago.threads.permissions import exclude_invisible_posts
+from misago.threads.serializers import DeletePostsSerializer
 
 
 DELETE_LIMIT = settings.MISAGO_POSTS_PER_PAGE + settings.MISAGO_POSTS_TAIL
@@ -27,12 +28,26 @@ def delete_post(request, thread, post):
 
 
 def delete_bulk(request, thread):
-    posts = clean_posts_for_delete(request, thread)
+    serializer = DeletePostsSerializer(
+        data={'posts': request.data},
+        context={
+            'thread': thread,
+            'user': request.user,
+        },
+    )
 
-    for post in posts:
+    if not serializer.is_valid():
+        if 'posts' in serializer.errors:
+            errors = serializer.errors['posts']
+        else:
+            errors = list(serializer.errors.values())[0]
+        return Response({'detail': errors[0]}, status=400)
+
+    for post in serializer.validated_data['posts']:
         post.delete()
 
     sync_related(thread)
+
     return Response({})
 
 
@@ -42,38 +57,3 @@ def sync_related(thread):
 
     thread.category.synchronize()
     thread.category.save()
-
-
-def clean_posts_for_delete(request, thread):
-    posts_ids = clean_ids_list(
-        request.data or [],
-        _("One or more post ids received were invalid."),
-    )
-
-    if not posts_ids:
-        raise PermissionDenied(_("You have to specify at least one post to delete."))
-    elif len(posts_ids) > DELETE_LIMIT:
-        message = ungettext(
-            "No more than %(limit)s post can be deleted at single time.",
-            "No more than %(limit)s posts can be deleted at single time.",
-            DELETE_LIMIT,
-        )
-        raise PermissionDenied(message % {'limit': DELETE_LIMIT})
-
-    posts_queryset = exclude_invisible_posts(request.user, thread.category, thread.post_set)
-    posts_queryset = posts_queryset.filter(id__in=posts_ids).order_by('id')
-
-    posts = []
-    for post in posts_queryset:
-        post.category = thread.category
-        post.thread = thread
-        if post.is_event:
-            allow_delete_event(request.user, post)
-        else:
-            allow_delete_post(request.user, post)
-        posts.append(post)
-
-    if len(posts) != len(posts_ids):
-        raise PermissionDenied(_("One or more posts to delete could not be found."))
-
-    return posts
